Removes unused aggregated metrics from materialized tables (Recidiviz/recidiviz-data#29464)
## Description of the change

As part of an effort to reduce materialization costs of aggregated
metrics views, this PR removes a subset of metrics that aren't regularly
used by product/analysts from the view collector. All of these metric
configurations still exist in `aggregated_metric_configurations` and can
be referenced in notebooks as a custom metric (or easily added back to
the materialized table as the need arises).

Seeking feedback on whether this feels like the right subset of metrics
to exclude.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of #Recidiviz/recidiviz-data#29291

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: cfff8328791156bed20955f9cabe8b7d30d110be