[Admin Panel] make sure new + updated users have the "roles" field (Recidiviz/recidiviz-data#30010)
## Description of the change

Noticed this bug when I tried to add a new user in staging and the page
crashed because the user had no "roles" field. Since we already did the
DB migration for this field, I figured it was better to make sure it
always gets added rather than returning it if not set. Since some of the
backend endpoints are already updated to work with the new field, I made
the fix in the frontend.

Tested locally by adding a new user, editing an existing user's role,
and editing information about an existing user but not touching the
role. All succeeded.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#27969

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 61485f21f49feb46a8f199b64c3dc605e68ee576