[US_CA] Update supervision_type for folks who are Deported or lacking OffenderGroup AND map Cat E to limited (Recidiviz/recidiviz-data#30005)
## Description of the change

We previously just used OffenderGroup to map supervision_type. However,
there are some periods (especially from the past, or for the first week
we receive data) during which OffenderGroup would be null. In the cases
when OffenderGroup is null, but SupervisionLevel is present, we can
assume the person is either on PAROLE or DEPORTED. This PR updates the
mapping to first use SupervisionLevel to see if someone is DEPORTED.
Then it checks OffenderGroup for absconsions, and finally if
OffenderGroup is null, we default to PAROLE if SupervisionLevel is
present (and not null).

This change also maps Category E to limited.

Tests were updated to reflect this change.

[Sandbox
passed](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-05-22_10_56_54-15757057238454078209;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))!

@brendali121 I was hoping you could weigh in with your Deep Insights
Insights on if you foresee any issues with this change. Specifically,
this change maps around 200k periods for which supervision_type was
previously null to now have a supervision_type of PAROLE or DEPORTED.
However, none of these are current periods:
```
-- returns no results
select count(*)
from `normalized_state.state_supervision_period`
where state_code = 'US_CA'
and termination_date is null
and supervision_type is null;
```
Some current folks who were Deported will have their supervision_type
correctly changed from PAROLE to DEPORTED.

I feel pretty confident these changes should be okay, but wanted to run
it through your brain to be sure!

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29969

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: eb024fdfa73ee78a88df3b473dda720b9a8fb201