[sftp] update public_key auth flow (Recidiviz/recidiviz-data#30082)
## Description of the change

i think in the past there was an exception we needed to catch during the
public key auth flow in order to send our password to pa. now it seems
it doesnt raise an error; instead, it kindly tells us what we need to
send next. See successful run for all sftp connections in test env here:
https://b4286596a08b49a2942b0b4edf6825a3-dot-us-central1.composer.googleusercontent.com/dags/recidiviz-staging_sftp_dag/grid?dag_run_id=manual__2024-05-24T16%3A12%3A53.256970%2B00%3A00&task_id=US_PA.find_sftp_files_to_download&tab=logs

Plan is to:
- merge this
- prep CP that includes Recidiviz/recidiviz-data#30062 and this PR to go out today to resolve
the issue in production as well

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#30044

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a0b0482fd8793e3205767a0c937b62a65e23a1b4