upgrade terraform google providers to 5.30.0 (Recidiviz/recidiviz-data#30024)
## Description of the change

This also required upgrading the `serverless_negs` module- I went ahead
and upgraded it all the way to 11.0 so hopefully it's a while before we
need to again.

The only two changes I had to make were:
1. adding a `cdn_policy` on our load balancers that had `enable_cdn =
true`, but I just used the existing values from the state file so there
are no diffs.
2. updating `automatic = true` to `auto {}` on the
google_secret_manager_secret resources

There are a lot of changes in the plan, but almost all of them are
adding `terraform_labels` as per
https://registry.terraform.io/providers/hashicorp/google/latest/docs/guides/version_5_upgrade#provider-level-labels-rework

Full plan output at
https://gist.github.com/danawillow/fd49a550b4ea6d0fbae3205934122fec.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Dan Hansen <dan@recidiviz.org>
GitOrigin-RevId: 3e431f1f4eb08c4e392b32a1b097dc8de3338f5f