[Workflows] remove locations-as-staff logic from BQ+ETL (Recidiviz/recidiviz-data#30066)
## Description of the change

When we started developing MO RH workflows, we stuck a location in the
staff record in order to develop the workflow faster, and then followed
up by introducing the concept of searchables and the location record.
Now that we have those, we can keep the incarceration staff record to
only be for people who we are actually showing in search results. Right
now, ME is the only state that does search-by-person for facilities
workflows.

This fixes the issues noted in the MI RH bug bash where facilities were
shown twice in search results, and also people were seeing "case
manager: [facility name]" in the sidebar.

Loaded the two changed views to sandbox at `dana_facility_staff`. The
resident record has more entries in it because we no longer filter out
cases where the officer ID is null, which seems reasonable to me. There
will be some people in the resident record that don't show up in search
results because of this, but I think it's worth it over having to
maintain the filter logic.

~~This relies on
https://github.com/Recidiviz/recidiviz-dashboards/issues/5396, which has
been deployed to staging in the FE but not prod yet. I feel pretty
comfortable merging this now since it won't be in prod until Tuesday,
and the FE should be deployed to prod before then. We can also just wait
until the FE has been deployed to prod 🤷‍♀️~~ jk it's in prod now!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29511

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4f59566c6d42b92ef217e4d4064075b6952b1bb3