[Workflows] Reflect Snooze Status in Opportunity Status Views (Recidiviz/recidiviz-data#30122)
## Description of the change

This is another try at Recidiviz/recidiviz-data#28633, which had to be reverted because it
relied on a table that didn't exist on staging. I've copied
`pulse_dashboard_segment_analytics.frontend_opportunity_snoozed` from
prod to staging and reconfigured
`workflows_views.clients_opportunity_snoozed` to deploy on staging.

The original PR description:

This PR adds a clients_snooze_spans view that captures the scheduled and
actual periods when a client or resident is snoozed in the Workflows
tool. Using that view, it adds a new DENIAL_ENDED state to
clients_latest_referral_status that clients enter when their snoozes
expire. It behaves the same as the NOT_VIEWED state in that any
subsequent user action will move them out of the state.

client_snooze_spans should now contain an entry for every period in
which a person has ever been marked as ineligible, even those that
occurred pre-snooze.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4fcfda791f6746f18df2dd92097bbf304560c02f