[Dataflow] Expose only StateFilteredQueryProvider for reference queries in base pipeline (Recidiviz/recidiviz-data#30154)
## Description of the change

I started looking into Recidiviz/recidiviz-data#29514 (read from `us_xx_state_primary` instead
of `state` for normalization pipelines) and realized it was quite a bit
trickier than I'd thought because we have reference view inputs to the
pipelines which also read from `state`. We need to be able to
dynamically generate those queries so that we can pick which state
dataset to query from.

This refactor changes the `BasePipeline` interface from:
```
def all_input_reference_view_builders(cls) -> List[BigQueryViewBuilder]:
```

to

```
    def all_input_reference_query_providers(
        cls, state_code: StateCode, address_overrides: BigQueryAddressOverrides | None
    ) -> Dict[str, StateFilteredQueryProvider]:
```

This gives individual pipelines the options to either back their
reference queries with `BigQueryView` objects (eventually should be
disallowed but the case for all pipelines right now) or other query
provider objects that are built from a simpler string template.

Eventually I will follow with a change that actually updates the
pipelines so they build the queries dynamically at runtime based on the
state code.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#29514 and Recidiviz/recidiviz-data#29518

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9726f7d84f1a404ffac26a2cd57add0a563da2d9