[Insights] Add duplicate for supervision_officers table (Recidiviz/recidiviz-data#30119)
## Description of the change

This PR adds the `supervision_officers` BQ view to the views to export
for Insights, while also adding the array of supervisor ids to said
view. I also added a duplicate schema definition in `insights/schema.py`
with the table name equivalent to the product view so that we can
hydrate this table in the insights instance.

The overall goal is to rename the existing table for the
SupervisionOfficer entity in insights/schema.py from
`insights_supervision_officers` to `supervision_officers` without
creating any product downtime. Once the new table in the insights
instance is hydrated, then I'll delete the duplicate schema definition
and swap the table name such that `supervision_officers` is used in
Recidiviz/recidiviz-data#29763.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: db14130198845f2a0df2761ad807ed3a97757f18